Created Flower Prefabs, Created HUD, Created PostProcessing profiles

+++ Assets/2024PegGameJam/Scripts/Interfaces/IDamageable.cs
@@ -4,5 +4,5 @@ using UnityEngine;
 
 public interface IDamageable
 {
-    void MakeDamage(float damage, GameObject instigator);
+    void MakeDamage(int damage, GameObject instigator);
 }
